Updated the Tile Class structure to make the MapCoords a Vector3Int. This makes dealing with Tiles and thier placements in Maps much simpler. Also threw in some changes to other files that were almost entirely visual and just to change the inspector UI. The exception is the Dialogue which only got an access change to its 'IsProcessing' member.

+++ Assets/_Dev Assets/Project Data Assets/WorldMaps/Builder/LinkedTileLocater.cs
@@ -33,7 +33,7 @@ public class LinkedTileLocater : MonoBehaviour
                 continue;
             }
 
-            if (Math.Abs((int)mTileOrigin.MapCoords.x - (int)mTile.MapCoords.x) == 1)
+            if (Math.Abs(mTileOrigin.MapCoords.x - mTile.MapCoords.x) == 1)
             {
                 if (CompareYZmTiles(mTileOrigin, mTile))
                 {
@@ -42,7 +42,7 @@ public class LinkedTileLocater : MonoBehaviour
                 }
             }
 
-            if (Math.Abs((int)mTileOrigin.MapCoords.z - (int)mTile.MapCoords.z) == 1)
+            if (Math.Abs(mTileOrigin.MapCoords.z - mTile.MapCoords.z) == 1)
             {
                 if (CompareYXmTiles(mTileOrigin, mTile))
                 {
